+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/FreightFrenzy_2021/gary/TeleOp_Intake_Spin.java
@@ -1,6 +1,5 @@
 package org.firstinspires.ftc.teamcode.FreightFrenzy_2021.gary;
 
-import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.hardware.DcMotor;
@@ -8,7 +7,6 @@ import com.qualcomm.robotcore.util.ElapsedTime;
 import com.qualcomm.robotcore.util.Range;
 
 @TeleOp(name = "GTeleOp_Intake_Spin", group = "Linear OpMode")
-@Disabled
 public class TeleOp_Intake_Spin extends LinearOpMode {
     private DcMotor Intake = null;
     private DcMotor Spin = null;
